lxc/profile: Add refresh option for copy
lxc profile copy command would refuse to update a profile if it
already exist by sending an error message to the user. There are
cases where we want to update the target profile anyway.
This change adds the --refresh tag which allow to update an
existing target profile with the details of the source one.

Signed-off-by: Yao Noel Achi <achi.noel@hotmail.com>